chore(YALB-86): abstracted page-args for reuse
