+++ components/05-page-examples/page-examples.stories.js
@@ -1,13 +1,9 @@
-import tokens from '@yalesites-org/tokens/build/json/tokens.json';
+import argTypes from '../04-page-layouts/page-args';
 
 import basicPageTwig from './basic-page.twig';
 
+import utilityNavData from '../03-organisms/menu/utility-nav/utility-nav.yml';
 import breadcrumbData from '../03-organisms/menu/breadcrumbs/breadcrumbs.yml';
-
-const borderThicknessOptions = Object.keys(tokens.border.thickness);
-const primaryNavPositions = Object.keys(tokens.layout['flex-position']);
-const siteHeaderThemeOptions = Object.keys(tokens['site-header-themes']);
-const siteFooterThemeOptions = Object.keys(tokens['site-footer-themes']);
 
 /**
  * Storybook Definition.
@@ -17,43 +13,7 @@ export default {
   parameters: {
     layout: 'fullscreen',
   },
-  argTypes: {
-    siteName: {
-      name: 'Site Name',
-      type: 'string',
-      defaultValue: 'Department of Chemistry',
-    },
-    headerBorderThickness: {
-      name: 'Header: Border thickness',
-      options: borderThicknessOptions,
-      type: 'select',
-      defaultValue: '8',
-    },
-    primaryNavPosition: {
-      name: 'Header: Primary nav position',
-      options: primaryNavPositions,
-      type: 'select',
-      defaultValue: 'right',
-    },
-    siteHeaderTheme: {
-      name: 'Header: Theme',
-      options: siteHeaderThemeOptions,
-      type: 'select',
-      defaultValue: 'white',
-    },
-    footerBorderThickness: {
-      name: 'Footer: Border thickness',
-      options: borderThicknessOptions,
-      type: 'select',
-      defaultValue: '8',
-    },
-    siteFooterTheme: {
-      name: 'Footer: Theme',
-      options: siteFooterThemeOptions,
-      type: 'select',
-      defaultValue: 'blue-yale',
-    },
-  },
+  argTypes,
 };
 
 export const BasicPage = ({
@@ -61,6 +21,8 @@ export const BasicPage = ({
   headerBorderThickness,
   primaryNavPosition,
   siteHeaderTheme,
+  utilityNavLinkContent,
+  utilityNavSearch,
   siteFooterTheme,
   footerBorderThickness,
 }) =>
@@ -71,5 +33,8 @@ export const BasicPage = ({
     site_header__theme: siteHeaderTheme,
     site_footer__border_thickness: footerBorderThickness,
     site_footer__theme: siteFooterTheme,
+    ...utilityNavData,
+    utility_nav__link__content: utilityNavLinkContent,
+    utility_nav__search: utilityNavSearch,
     ...breadcrumbData,
   });
